[refactor/#147] develop merge
