+++ app/src/main/java/com/poti/android/core/common/extension/ModifierExt.kt
@@ -1,13 +1,10 @@
 package com.poti.android.core.common.extension
 
-import android.annotation.SuppressLint
 import android.graphics.BlurMaskFilter
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -24,13 +21,9 @@ import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.drawOutline
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-
-private val figmaScreenWidth = 360.dp
-private val figmaScreenHeight = 800.dp
 
 inline fun Modifier.noRippleClickable(
     crossinline onClick: () -> Unit,
@@ -164,20 +157,4 @@ fun Modifier.bottomBorder(
             strokeWidth = strokeWidthPx,
         )
     }
-}
-
-@SuppressLint("ConfigurationScreenWidthHeight")
-@Composable
-fun Modifier.heightForScreenPercentage(height: Dp): Modifier {
-    val screenHeight = LocalConfiguration.current.screenHeightDp.dp
-    val ratio = screenHeight / figmaScreenHeight
-    return this.height(height * ratio)
-}
-
-@SuppressLint("ConfigurationScreenWidthHeight")
-@Composable
-fun Modifier.widthForScreenPercentage(width: Dp): Modifier {
-    val screenWidth = LocalConfiguration.current.screenWidthDp.dp
-    val ratio = screenWidth / figmaScreenWidth
-    return this.width(width * ratio)
 }
