+++ app/src/main/java/com/poti/android/core/designsystem/component/modal/PotiLargeModal.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -16,7 +17,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.poti.android.R
-import com.poti.android.core.common.extension.heightForScreenPercentage
 import com.poti.android.core.designsystem.component.button.ModalButtonType
 import com.poti.android.core.designsystem.component.button.PotiModalButton
 import com.poti.android.core.designsystem.theme.PotiTheme
@@ -92,7 +92,7 @@ fun PotiLargeModal(
                     contentDescription = null,
                     modifier = Modifier
                         .padding(vertical = 24.dp)
-                        .heightForScreenPercentage(180.dp)
+                        .height(180.dp)
                         .aspectRatio(1f),
                     contentScale = ContentScale.Crop,
                 )
